Massive refactor\baseline
Revamped code structure and made everything run in a single module. No
longer dependant on INpureCore or API modules (An API will be released
in the future however).

Some design decisions re-evaluated. Carpenter is out, having a special
machine for 2 or 3 recipes was too niche. Integration with WAILA and NEI
to come also.